Added error in case of illegal type
